+++ syconn/reps/connectivity_helper.py
@@ -114,29 +114,29 @@ def load_cached_data_dict(thresh_syn_prob=None, axodend_only=True, wd=None,
         csd.load_numpy_data('syn_prob')
     # -1 for inhibitory, +1 for excitatory
     cd_dict['syn_sign'] = \
-        csd.load_numpy_data('syn_sign').astype(np.int)
+        csd.load_numpy_data('syn_sign').astype(np.int32)
     cd_dict['coord_x'] = \
-        csd.load_numpy_data('rep_coord')[:, 0].astype(np.int)
+        csd.load_numpy_data('rep_coord')[:, 0].astype(np.int32)
     cd_dict['coord_y'] = \
-        csd.load_numpy_data('rep_coord')[:, 1].astype(np.int)
+        csd.load_numpy_data('rep_coord')[:, 1].astype(np.int32)
     cd_dict['coord_z'] = \
-        csd.load_numpy_data('rep_coord')[:, 2].astype(np.int)
+        csd.load_numpy_data('rep_coord')[:, 2].astype(np.int32)
     cd_dict['ssv_partner_0'] = \
-        csd.load_numpy_data('neuron_partners')[:, 0].astype(np.int)
+        csd.load_numpy_data('neuron_partners')[:, 0].astype(np.int32)
     cd_dict['ssv_partner_1'] = \
-        csd.load_numpy_data('neuron_partners')[:, 1].astype(np.int)
+        csd.load_numpy_data('neuron_partners')[:, 1].astype(np.int32)
     cd_dict['neuron_partner_ax_0'] = \
-        csd.load_numpy_data('partner_axoness')[:, 0].astype(np.int)
+        csd.load_numpy_data('partner_axoness')[:, 0].astype(np.int32)
     cd_dict['neuron_partner_ax_1'] = \
-        csd.load_numpy_data('partner_axoness')[:, 1].astype(np.int)
+        csd.load_numpy_data('partner_axoness')[:, 1].astype(np.int32)
     cd_dict['neuron_partner_ct_0'] = \
-        csd.load_numpy_data('partner_celltypes')[:, 0].astype(np.int)
+        csd.load_numpy_data('partner_celltypes')[:, 0].astype(np.int32)
     cd_dict['neuron_partner_ct_1'] = \
-        csd.load_numpy_data('partner_celltypes')[:, 1].astype(np.int)
+        csd.load_numpy_data('partner_celltypes')[:, 1].astype(np.int32)
     cd_dict['neuron_partner_sp_0'] = \
-        csd.load_numpy_data('partner_spiness')[:, 0].astype(np.int)
+        csd.load_numpy_data('partner_spiness')[:, 0].astype(np.int32)
     cd_dict['neuron_partner_sp_1'] = \
-        csd.load_numpy_data('partner_spiness')[:, 1].astype(np.int)
+        csd.load_numpy_data('partner_spiness')[:, 1].astype(np.int32)
 
     log_reps.debug('Getting {1} objects took: {0}'.format(time.time() - start,
                                                           len(csd.ids)))
@@ -309,7 +309,7 @@ def plot_wiring(path, wiring, den_borders, ax_borders, cumul=False, log: Optiona
 
     # TODO: becomes slow for large entry_width
     bin_intensity_plot = intensity_plot != 0
-    bin_intensity_plot = bin_intensity_plot.astype(np.float)
+    bin_intensity_plot = bin_intensity_plot.astype(np.float32)
     intensity_plot = ndimage.convolve(intensity_plot, np.ones((entry_width, entry_width)))
     bin_intensity_plot = ndimage.convolve(bin_intensity_plot, np.ones((entry_width, entry_width)))
     intensity_plot /= bin_intensity_plot
